refactor(cli): remove baml grep
