+++ baml_language/crates/baml_lsp2_actions/src/describe_tests.rs
@@ -230,9 +230,10 @@ fn user_only_describe_still_does_not_search_builtins() {
 }
 
 #[test]
-fn list_symbols_compiler2_visible_includes_selected_builtins() {
+fn search_symbols_compiler2_visible_includes_selected_builtins() {
     let project = make_project();
-    let symbols = project.list_symbols_compiler2_visible();
+    let files = baml_compiler2_hir::compiler2_all_files(&project.db);
+    let symbols = crate::search::search_symbols(&project.db, &files, "");
 
     assert!(symbols.iter().any(|sym| sym.name == "String"));
     assert!(symbols.iter().any(|sym| sym.name == "Array"));
